# Conflicts:
#	sod/sod-dm/sod-dm-dao/src/main/java/com/piesat/dm/dao/datatable/DataTableDao.java
#	sod/sod-schedule-center/sod-schedule-rpc-service/src/main/java/com/piesat/schedule/rpc/service/statistics/TableCollectHandler.java

+++ sod/sod-ucenter/sod-ucenter-rpc-service/src/main/java/com/piesat/ucenter/rpc/service/system/UserServiceImpl.java
@@ -622,10 +622,12 @@ public class UserServiceImpl extends BaseService<UserEntity> implements UserServ
     @Transactional
     public void delBizUser(String bizUserid) {
         UserEntity userEntity = userDao.findByUserName(bizUserid);
-        userRoleDao.deleteByUserId(userEntity.getId());
-        this.delete(userEntity.getId());
+        if(userEntity != null){
+            userRoleDao.deleteByUserId(userEntity.getId());
+            this.delete(userEntity.getId());
 
-        //删除业务用户所关联的数据
-        dataClassService.deleteByBizUserId(bizUserid);
+            //删除业务用户所关联的数据
+            dataClassService.deleteByBizUserId(bizUserid);
+        }
     }
 }
